more layout improvements
